Added stuff

+++ src/daseel/game/trialsofjorah/Vector.java
@@ -5,10 +5,10 @@ package daseel.game.trialsofjorah;
  */
 public class Vector {
 
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 
-	public Vector(int x, int y) {
+	public Vector(float x, float y) {
 
 		this.x = x;
 		this.y = y;
@@ -19,21 +19,21 @@ public class Vector {
 		this(vector.getX(), vector.getY());
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public int getY() {
+	public float getY() {
 
 		return y;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 
 		this.x = x;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 
 		this.y = y;
 	}
